Small innocent stuff

+++ drivers/gpio.c
@@ -10,13 +10,13 @@
 #include "boards.h"
 
 
-void gpio_set(uint16_t gpio_nb_) {
+void gpio_set(uint32_t gpio_nb_) {
 
 	nrf_gpio_pin_set(gpio_nb_);
 
 }
 
-bool gpio_get(uint16_t gpio_nb_) {
+bool gpio_get(uint32_t gpio_nb_) {
 
 	return nrf_gpio_pin_read(gpio_nb_);
